Why cast a constant to (unsigned long) when you can make it this by appending UL.
Former-commit-id: 85f00b679888fcf2fd685d1008febbdf40443132

+++ usr.sbin/timed/timedc/cmds.c
@@ -239,7 +239,7 @@ begin_define
 define|#
 directive|define
 name|BU
-value|((unsigned long)2208988800)
+value|(2208988800UL)
 end_define
 
 begin_comment
